1.地图点击优化 2.统计分析 3.全量表格

+++ modules/tables/tables.extraQuery.js
@@ -21,23 +21,29 @@ define("tables/extraQuery", [
         _timestamp: null,
         _data: [],
         _hash: {},
-        __fieldAliases: {},
+        _fieldAliases: {},
         _force: null,
         _config: null,
-        _index:null,
+        _index: null,
+        _titleConfig: {
+            1954: { k: 'LDGS', t: '所属公园' },
+            1960: { k: 'LDGS', t: '绿地归属' },
+            1947: { k: 'NAME', t: '古树名称' },
+            1983: { k: 'NAME', t: '名称' },
+        },
         /**
          * 表模板
          */
         _template: `<div class='extra_obj extra_query extra_query_###'><div>
-            <h4>数据搜索</h4>
+            <h4>数据搜索<button class='btn btn-primary btn-updown' data-info="on">收起</button></h4>
                <div class="extra_header form-inline">
                 <div class="form-group mb-3"><label>类型: </label><select class="form-control extra_query_select">
-                    <option value='1954@LDGS' selected >公园绿地</option>
-                    <option value='1960@LDGS'>道路绿地</option>
-                    <option value='1947@NAME'>古木名树</option>
-                    <option value='1983@NAME'>地名地址</option>
+                    <option value='1954' selected >公园绿地</option>
+                    <option value='1960'>道路绿地</option>
+                    <option value='1947'>古木名树</option>
+                    <option value='1983'>地名地址</option>
                 </select></div>
-                <div class="form-group mb-3"><label>名称搜索: </label><input class="form-control searchName"/></div>
+                <div class="form-group mb-3"><label class="extra_query_title">所属公园: </label><input class="form-control searchName"/></div>
                 <button class='btn btn-primary extra_query_search'>搜索</button>
                 <button class='btn btn-primary extra_query_export'>导出</button>
                 <span class="extra_query_close">x<span>
@@ -49,7 +55,7 @@ define("tables/extraQuery", [
         /**
          * 过滤字段
          */
-        _banned: ["FEATUREGUID", "PICTURE"],
+        _banned: ["OBJECTID", "OBJECTID_1", "FEATUREGUID", "PICTURE"],
         initialize: function (id = 'extra_querys') {
             //  this._table = $(`#${id}`).DataTables();
             this._id = id;
@@ -66,8 +72,8 @@ define("tables/extraQuery", [
         queryTable: function () {
             const that = this;
             const arcgisxhr = new L.DCI.ArcgisXhr();
-            const [id, index] = ($(".extra_query_select").val() || '1954@LDGS').split('@');
-            this._index = index;
+            const id = ($(".extra_query_select").val() || '1954').split('@');
+            this._index = that._titleConfig[id].k;
             const searchName = $(".searchName").val().trim();
             L.dci.app.services.baseService.getFeatureLayerById({
                 id,
@@ -79,7 +85,7 @@ define("tables/extraQuery", [
                         const { Url, LayerIndex } = that._config;
                         arcgisxhr.getArcgisByXhr(`${Url}/${LayerIndex}/query`, ({ fieldAliases, features }) => {
                             const _hash = {};
-                            const data = features.map((v, index) => {
+                            const data = features.map(v => {
                                 _hash[v.attributes.OBJECTID] = v;
                                 return v.attributes;
                             });
@@ -94,7 +100,7 @@ define("tables/extraQuery", [
                             } else {
                                 that.initDetailTableData(data);
                             }
-                        }, !searchName ? "1=1" : encodeURIComponent(`${index} like '%${searchName}%'`));
+                        }, !searchName ? "1=1" : encodeURIComponent(`${that._index} like '%${searchName}%'`));
                     } else {
                         L.dci.app.util.dialog.alert("提示", "该服务已被禁用");
                     }
@@ -106,7 +112,8 @@ define("tables/extraQuery", [
                 serverSide: false,
                 data,
                 ordering: true,
-                scrollY: 160,
+                scrollY: 340,
+                pageLength: 20,
                 scrollX: true,
                 bSort: true,
                 searching: false,
@@ -152,38 +159,53 @@ define("tables/extraQuery", [
             //  [表格]   搜索
             $('body').on('click', `.extra_query_${that._timestamp} .extra_query_search`, function () {
                 that.queryTable();
+            });
+            //  [表格] 收起&展开
+            $('body').on('click', `.extra_query_${that._timestamp} .btn-updown`, function () {
+                const ison = $(this).attr("data-info") == "on";
+                ison && $(".extra_obj").addClass("extra_obj_hidden") && $(this).attr("data-info", "off").text("展开");
+                !ison && $(".extra_obj").removeClass("extra_obj_hidden") && $(this).attr("data-info", "on").text("收起");
             })
+            //  [表格] 类型修改
+            $('body').on('change', `.extra_query_${that._timestamp} .extra_query_select`, function () {
+                $(".extra_query_title").text(that._titleConfig[$(".extra_query_select").val()].t);
+                
+            });
             //  [表格] 点击条目
             $('body').on('click', `.extra_query_${that._timestamp} tbody tr`, function () {
                 const data = that._hash[that._table.fnGetData(this).OBJECTID];
                 //  赋值
                 that._force = data.attributes;
                 //  定位
-                const { type, geometry } = that.doTransform(data.geometry);
-                const map = L.DCI.App.pool.get('MultiMap').getActiveMap();
-                const hlLayer = map.getLabelLayer();
-                hlLayer.clearLayers();
-                if (type == 'point') {
-                    L.marker(geometry).addTo(hlLayer);
-                    L.popup({ maxWidth: 80, offsety: 10, className: 'popupLittleUp' })
-                        .setLatLng(geometry)
-                        .setContent(`${data.attributes.ADDRESS} - [${data.attributes[that._index]}]`)
-                        .openOn(map.map);
-                    map.map.panTo(geometry);
-                } else if (type == 'polygon') {
-                    const polygon = L.polygon(geometry, { color: 'red' }).addTo(hlLayer);
-                    console.log(polygon.getBounds());
-                    map.map.fitBounds(polygon.getBounds());
-                }
+                const arcgisxhr = new L.DCI.ArcgisXhr();
+                const { Url, LayerIndex } = that._config;
+                arcgisxhr.getArcgisByXhr(`${Url}/${LayerIndex}/query`, ({ features }) => {
+                    const { type, geometry } = that.doTransform(features[0].geometry);
+                    const map = L.DCI.App.pool.get('MultiMap').getActiveMap();
+                    const hlLayer = map.getLabelLayer();
+                    hlLayer.clearLayers();
+                    if (type == 'point') {
+                        L.marker(geometry).addTo(hlLayer);
+                        L.popup({ maxWidth: 80, offsety: 10, className: 'popupLittleUp' })
+                            .setLatLng(geometry)
+                            .setContent(data.attributes[that._index])
+                            .openOn(map.map);
+                        map.map.panTo(geometry);
+                    } else if (type == 'polygon') {
+                        const polygon = L.polygon(geometry, { color: 'red' }).addTo(hlLayer);
+                        map.map.fitBounds(polygon.getBounds());
+                    }
 
-                //  弹框
-                $(".extra_details").remove();
-                $("body").append(extra);
-                that.doBasicDisplay();
+                    //  弹框
+                    $(".extra_details").remove();
+                    $("body").append(extra);
+                    $(".extra_obj").addClass("extra_obj_hidden") && $(".btn-updown").attr("data-info", "off").text("展开")
+                    that.doExtraEvent();
+                }, `OBJECTID=${data.attributes.OBJECTID}`, true);
             });
             //  [表格]   点击退出
             $('body').on('click', `.extra_query_${that._timestamp} .extra_query_close`, function () {
-                that._table.api().destroy();
+                that._table && that._table.api().destroy();
                 $(".extra_query").remove();
             })
             //  [表格]   点击导出
@@ -193,12 +215,16 @@ define("tables/extraQuery", [
                     return v.attributes;
                 }))
             })
+        },
+        doExtraEvent: function () {
+            const that = this;
+            $(".extra_details").unbind();
             //  [弹出框] 点击退出
-            $('body').on('click', '.extra_close', function () {
+            $('.extra_details').on('click', '.extra_close', function () {
                 $('.extra_details').remove();
             })
             //  [弹出框] 切换
-            $('body').on('click', '.extra_details_body_left_tip', function () {
+            $('.extra_details').on('click', '.extra_details_body_left_tip', function () {
                 const type = $(this).attr('data-info');
                 switch (type) {
                     case 'basic': {
@@ -215,6 +241,7 @@ define("tables/extraQuery", [
                     }
                 }
             })
+            that.doBasicDisplay();
         },
         doBasicDisplay: function () {
             $('.extra_details_body_main').html(`<div><header>基本信息</header><ul>${Object.keys(this._force).filter(v => !~this._banned.indexOf(v)).map(v => { return `<li><label>${this._fieldAliases[v]}</label><span>${this._force[v]}</span></li>` }).join('')}</ul></div>`);
@@ -226,6 +253,7 @@ define("tables/extraQuery", [
             $('.extra_details_body_main').html(`<div><header>图片展示</header></div>`);
         }, 
         tableDefault: function () {
+            $.fn.dataTable.defaults.fnFormatNumber = function (v) { return v };
             $.fn.dataTable.defaults.oLanguage = {
                 "sProcessing": "处理中...",
                 "sLengthMenu": "显示 _MENU_ 项结果",
@@ -252,6 +280,27 @@ define("tables/extraQuery", [
             };
         },
         exportTable: function (jsonData, worksheet = +new Date()) {
+            const str = `${Object.keys(jsonData[0])
+                .map((v) => `${this._fieldAliases[v] || ``},`)
+                .join(``)}\n${jsonData
+                    .map(
+                        (v) =>
+                            `${Object.keys(v)
+                                .map((d) => `${v[d] || ``}`)
+                                .join(`,`)}`
+                    )
+                    .join(`\n`)}`;
+            var link = document.createElement("a");
+            var csvContent = "data:text/csv;charset=utf-8,\uFEFF" + str;
+            var encodedUri = csvContent;
+            console.log(encodedUri.length)
+            link.setAttribute("href", encodedUri);
+            link.setAttribute("download", `${this._config.LayerName}-${worksheet}.csv`);
+            document.body.appendChild(link);
+            link.click();
+            document.body.removeChild(link);
+        },
+        exportTable_: function (jsonData, worksheet = +new Date()) {
             const excelContent = `<tr>${Object.keys(jsonData[0])
                 .map((v) => `<td>${this._fieldAliases[v] || ``}</td>`)
                 .join(``)}</tr>${jsonData
@@ -274,9 +323,9 @@ define("tables/extraQuery", [
             //下载模板
             // window.location.href = ;
             const link = document.createElement("a");
-            link.href = uri + encodeURIComponent(window.btoa(unescape(encodeURIComponent(template))));
+            link.href = uri + window.btoa(unescape(encodeURIComponent(template)));
             //对下载的文件命名
-            link.download = `${worksheet}.xls`;
+            link.download = `${this._config.LayerName}-${worksheet}.xls`;
             document.body.appendChild(link);
             link.click();
             document.body.removeChild(link);
